Fix block listeners for interactions
