+++ spock-core/src/main/java/org/spockframework/compiler/SpecRewriter.java
@@ -22,6 +22,7 @@ import org.spockframework.util.*;
 
 import java.lang.reflect.InvocationTargetException;
 import java.util.*;
+import java.util.stream.Collectors;
 
 import org.codehaus.groovy.ast.*;
 import org.codehaus.groovy.ast.expr.*;
@@ -394,8 +395,11 @@ public class SpecRewriter extends AbstractSpecVisitor implements IRewriteResourc
     this.block = null;
 
     if (!movedStatsBackToMethod)
-      for (Block b : method.getBlocks())
+      for (Block b : method.getBlocks()) {
+        // this will only have the blocks if there was no 'cleanup' block in the method
+        addBlockListeners(b);
         method.getStatements().addAll(b.getAst());
+      }
 
     // for global required interactions
     if (method instanceof FeatureMethod)
@@ -410,6 +414,59 @@ public class SpecRewriter extends AbstractSpecVisitor implements IRewriteResourc
 
     if (method instanceof FeatureMethod)
       clearCurrentBlockOnExit(method.getStatements());
+  }
+
+
+  private void addBlockListeners(Block block) {
+    BlockParseInfo blockType = block.getParseInfo();
+    if (blockType == BlockParseInfo.WHERE
+      || blockType == BlockParseInfo.METHOD_END
+      || blockType == BlockParseInfo.ANONYMOUS) return;
+
+    // SpockRuntime.enterBlock(getSpecificationContext(), new BlockInfo(blockKind, [blockTexts]))
+    MethodCallExpression enterBlockCall = createBlockListenerCall(block, blockType, nodeCache.SpockRuntime_CallEnterBlock);
+    // SpockRuntime.exitedBlock(getSpecificationContext(), new BlockInfo(blockKind, [blockTexts]))
+    MethodCallExpression exitBlockCall = createBlockListenerCall(block, blockType, nodeCache.SpockRuntime_CallExitBlock);
+
+    // As the cleanup block finalizes the specification, it would override any previous block in ErrorInfo,
+    // so we only call enterBlock if there is no error yet.
+    if (blockType == BlockParseInfo.CLEANUP) {
+      block.getAst().add(0, ifThrowableIsNull(enterBlockCall));
+      block.getAst().add(ifThrowableIsNull(exitBlockCall));
+    } else {
+      block.getAst().add(0, new ExpressionStatement(enterBlockCall));
+      block.getAst().add( new ExpressionStatement(exitBlockCall));
+    }
+  }
+
+  private IfStatement ifThrowableIsNull(MethodCallExpression methodCall) {
+    return new IfStatement(
+      // if ($spock_feature_throwable == null)
+      new BooleanExpression(AstUtil.createVariableIsNullExpression(new VariableExpression(SpecRewriter.SPOCK_FEATURE_THROWABLE, nodeCache.Throwable))),
+      new ExpressionStatement(methodCall),
+      EmptyStatement.INSTANCE
+    );
+  }
+
+  private MethodCallExpression createBlockListenerCall(Block block, BlockParseInfo blockType, MethodNode blockListenerMethod) {
+    return createDirectMethodCall(
+      new ClassExpression(nodeCache.SpockRuntime),
+      blockListenerMethod,
+      new ArgumentListExpression(
+        createDirectMethodCall(VariableExpression.THIS_EXPRESSION,
+          nodeCache.SpecInternals_GetSpecificationContext,
+          ArgumentListExpression.EMPTY_ARGUMENTS),
+        new ConstructorCallExpression(nodeCache.BlockInfo,
+          new ArgumentListExpression(
+            new PropertyExpression(
+              new ClassExpression(nodeCache.BlockKind),
+              blockType.name()
+            ),
+            new ListExpression(
+              block.getDescriptions().stream().map(ConstantExpression::new).collect(Collectors.toList())
+            )
+          ))
+      ));
   }
 
   @Override
@@ -490,6 +547,9 @@ public class SpecRewriter extends AbstractSpecVisitor implements IRewriteResourc
   @Override
   public void visitCleanupBlock(CleanupBlock block) {
     for (Block b : method.getBlocks()) {
+      // call addBlockListeners() here, as this method will consume the blocks,
+      // so we need to transform here as they will be gone in visitMethodAgain where we normally add them
+      addBlockListeners(b);
       if (b == block) break;
       moveVariableDeclarations(b.getAst(), method.getStatements());
     }
